Fix colors, unpack presets
There were two problems that were messing up the colors
- the tint method in quil had a bug in the rgba arity
- quil no longer allows passing a `color` to `tint`

Tint is also stateful, making it easy for an earlier setting to "bleed" into
other parts of the code.

This introduces a `with-fill` macro to keep the fill setting local.

When running not directly from source, MiniBeast crashes cause it can't find the
presets. Unpack them on first run.

Allow passing presets on the CLI